Factor out complex methods; fix some eslint issues
