+++ app/javascript/PCSG/RunDimensions.js
@@ -10,7 +10,14 @@ class RunDimensions {
   lanePlacement: ?number;
   laneSpan: ?number;
 
-  constructor(eventRun: EventRun, laneIndex: number, timePlacement: number, timeSpan: number, lanePlacement: ?number, laneSpan: ?number) {
+  constructor(
+    eventRun: EventRun,
+    laneIndex: number,
+    timePlacement: number,
+    timeSpan: number,
+    lanePlacement: ?number,
+    laneSpan: ?number,
+  ) {
     this.eventRun = eventRun;
     this.lanePlacement = lanePlacement;
     this.laneSpan = laneSpan;
